0_api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
PyYAML==6.0.2
docker==7.1.0